// types.ts
export type ClassLevel = '6' | '7' | '8' | '9' | '10' | '11' | '12' | 'COMPETITION';

export type Board = 'CBSE' | 'BSEB' | 'COMPETITION';

export type Stream = 'Science' | 'Commerce' | 'Arts';

export type Role = 'STUDENT' | 'ADMIN' | 'SUB_ADMIN';

export interface TimeConfig {
    days: number;
    hours: number;
    minutes: number;
    seconds: number;
}

export interface FrequencyConfig {
    value: number;
    unit: 'seconds' | 'minutes' | 'hours' | 'days' | 'months' | 'years';
}

export type ContentType = 'NOTES_SIMPLE' | 'NOTES_PREMIUM' | 'MCQ_ANALYSIS' | 'MCQ_SIMPLE' | 'PDF_FREE' | 'PDF_PREMIUM' | 'PDF_ULTRA' | 'PDF_VIEWER' | 'WEEKLY_TEST' | 'VIDEO_LECTURE' | 'NOTES_HTML_FREE' | 'NOTES_HTML_PREMIUM' | 'NOTES_IMAGE_AI';
// ...
export interface Subject {
  id: string;
  name: string;
  icon: string;
  color: string;
}

export interface Chapter {
  id: string;
  title: string;
  description?: string;
}